+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Don't convert to multi-part geometries by default in `copy_layer`,... (#570)
 - Improve performance of erase for very complex input geometries (#585)
 - Add configuration option to only warn on dissolve errors (#561)
+- Add support for renaming layer with only difference in casing (#593)
 
 ## 0.9.1 (2024-07-18)